pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
authors = [
    {name = "Aleksey", email = "pdskate88@gmail.com"},
]
readme = "README.md"
repository = "https://github.com/Habited/python-project-50"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "pytest-cov>=6.0.0",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["difference_calculator"]


[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.6",
]

[project.scripts]
gendiff = "difference_calculator.scripts.main:main"
